Full resolution of inf + clarification of NaNs.
Half precision infinities are now correctly handled and follow
the semantics from numpy. This commit also clarifies the situation with
NaNs. We do not accept NaNs as valid input to the table, which
simplifies  the handling of missing values. It may be possible to
support NaNs as input and to keep the missing value semantics, but it's
problematics. This is at least correct.