Integrated Numpy half float code successfully.
The numpy code is preferable to the half code for several reasons.
Firstly, it is part of a respected and maintained project, so likely
to be well tested. It is also much easier to read and follow.
Numpy is licensed under a BSD license, so there should be no issue
with including it here.